State with weather, API call, render weather component
Initialized state inside of the React App.js component to retain weather
info after making an API call.

I also made an API call to the same endpoint but one is commented out.
In order to serve as an example of how to make API calls with
async/await and .then().catch()

We also crated a nullcheck function to handle the null state at render
time and avoid an aerror being thrown